+++ tomography_preprocessing/tilt_series_alignment/imod_etomo_launcher.py
@@ -2,6 +2,7 @@ from pathlib import Path
 
 import starfile
 
+import tomography_preprocessing.tilt_series_alignment._utils
 from ._cli import cli
 from ..utils.relion import relion_pipeline_job
 from .. import utils
@@ -13,7 +14,7 @@ def etomo_launcher(
         output_directory: Path,
 ):
     tilt_series_directory, imod_alignments_directory = \
-        utils.imod.create_job_directory_structure(output_directory)
+        tomography_preprocessing.tilt_series_alignment._utils.create_alignment_job_directory_structure(output_directory)
     tilt_series_df = starfile.read(tilt_series_star_file, always_dict=True)['global']
 
 
@@ -23,5 +24,5 @@ class EtomoLauncher:
             self, tilt_series_star_file: Path, output_directory: Path
     ):
         self.tilt_series_directory, self.imod_alignments_directory = \
-            utils.imod.create_job_directory_structure(output_directory)
+            tomography_preprocessing.tilt_series_alignment._utils.create_alignment_job_directory_structure(output_directory)
         self.tilt_series_df = starfile.read(tilt_series_star_file, always_dict=True)['global']
