+++ tomography_preprocessing/utils/imod.py
@@ -1,14 +1,14 @@
 import os
 from functools import lru_cache
 from pathlib import Path
-from typing import Tuple, Callable, Dict, Any
+from typing import Tuple
 
 import numpy as np
 import pandas as pd
 import starfile
 
 from .. import utils
-from ..utils.transformations import S, Rx, Ry, Rz
+from ..utils.transformations import S, Ry, Rz
 
 
 def read_xf(file: os.PathLike) -> np.ndarray:
@@ -89,7 +89,7 @@ def get_tilt_series_alignment_parameters(
     """Get the tilt-series alignment parameters from an IMOD directory.
 
     Shifts are in pixels and should be applied before rotations.
-    Rotations are ZYZ intrinsic Euler angles which transform the volume
+    Rotations are ZYZ intrinsic Euler angles which transform the volume.
     """
     xf = read_xf(get_xf_file(imod_directory))
     shifts_px = get_pre_rotation_shifts(xf)
@@ -142,62 +142,6 @@ def relion_tilt_series_alignment_parameters_to_relion_matrix(
     volume_transforms = in_plane_rotation @ tilt @ tomogram_origin_offset
     image_transforms = tilt_image_corner_to_center @ tilt_image_centering_shift
     return np.squeeze(image_transforms @ volume_transforms)
-
-
-def create_job_directory_structure(output_directory: Path) -> Tuple[Path, Path]:
-    """Create directory structure for an IMOD alignment job."""
-    tilt_series_directory = output_directory / 'tilt_series'
-    tilt_series_directory.mkdir(parents=True, exist_ok=True)
-
-    imod_alignments_directory = output_directory / 'alignments'
-    imod_alignments_directory.mkdir(parents=True, exist_ok=True)
-    return tilt_series_directory, imod_alignments_directory
-
-
-def align_single_tilt_series(
-        tilt_series_id: str,
-        tilt_series_df: pd.DataFrame,
-        tilt_image_df: pd.DataFrame,
-        alignment_function: Callable,
-        alignment_function_kwargs: Dict[str, Any],
-        output_directory: Path,
-):
-    # Create output directory structure
-    tilt_series_directory, imod_alignments_directory = \
-        create_job_directory_structure(output_directory)
-    imod_directory = imod_alignments_directory / tilt_series_id
-    imod_directory.mkdir(parents=True, exist_ok=True)
-
-    # Establish filenames
-    tilt_series_filename = f'{tilt_series_id}.mrc'
-    tilt_image_metadata_filename = f'{tilt_series_id}.star'
-    tilt_series_path = tilt_series_directory / tilt_series_filename
-    tilt_image_metadata_star_path = tilt_series_directory / tilt_image_metadata_filename
-
-    # Order is important in IMOD, sort by tilt angle
-    tilt_image_df = tilt_image_df.sort_values(by='rlnTomoNominalStageTiltAngle', ascending=True)
-
-    # Create tilt-series stack and align using IMOD
-    # implicit assumption - one tilt-axis angle per tilt-series
-    utils.image.stack_image_files(
-        image_files=tilt_image_df['rlnMicrographName'],
-        output_image_file=tilt_series_path
-    )
-    alignment_function(
-        tilt_series_file=tilt_series_path,
-        tilt_angles=tilt_image_df['rlnTomoNominalStageTiltAngle'],
-        pixel_size=tilt_series_df['rlnMicrographOriginalPixelSize'],
-        nominal_rotation_angle=tilt_image_df['rlnTomoNominalTiltAxisAngle'][0],
-        output_directory=imod_directory,
-        **alignment_function_kwargs
-    )
-    utils.imod.write_relion_tilt_series_alignment_output(
-        tilt_image_df=tilt_image_df,
-        tilt_series_id=tilt_series_id,
-        pixel_size=tilt_series_df['rlnMicrographOriginalPixelSize'],
-        imod_directory=imod_directory,
-        output_star_file=tilt_image_metadata_star_path
-    )
 
 
 def write_aligned_tilt_series_star_file(
